Last pass at decoupling tests
